Several QOL improvments by Acta
Revert "Creates /sxos directory for SXOS users who did not have it"

This reverts commit bc1cb45c7a9c3cdf578a9ebeddbafb37d2eb85b4.

Title ID's that are found are now listed prior to downloading. Updated wording to more accurately reflect procedure.

Instread of saying it is downloading before it is certain, it informs the user of the most recent version, and then if it is not up to date, tells them it is downloading the latest.

Added functionality to get title names, which is not 100% working at the moment. Users can now return to the main menu at any time by pressing [-].

Rewrote the code to use vectors of tuples containing title information.

Exclusions not yet tested, and considering rewriting the vectors to take Title objects instead.

Added a title struct for future use.

Implemented Title struct test.

Added the ability to view which titles were updated after updating.

This needs to be expanded to ask for input beforehand.

Implemented a titles.hpp file that performs actions on the udpated.dat file, such as writing, reading and outputting.

Viewing updated titles is now optional.

Changed the screen so that it clears when a function is run, or when returning to the main menu.

Uses the consoleClear() function from console.h in libnx

Added a clear console function of my own, initially to reset the print window, but that function is not currently working.

Added confirmation screen before deleting cheat files.

Added some commenting

Added commenting